Added a new file that contains a script performing yearly projected savings calculation in python

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 #!/bin/bash
 #Calculates user's yearly potential savings based on monthly income and expenses and 5% interest rate
-income = int(input("Enter your monthly income: "))
-expenses = int(input("Enter your total monthly expenses: "))
+income = float(input("Enter your monthly income: "))
+expenses = float(input("Enter your total monthly expenses: "))
 
 savings = income - expenses
 projected_savings = savings * 12 + (savings * 12 * 0.05)
